requirements.txt
pdfservices-sdk
boto3
python-dotenv
